pyproject.toml
[tool.poetry]
name = "poetry_test"
version = "0.1.0"
description = ""
authors = ["mingyu <rlaalsrbgk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "<3"
selenium = "^3.141.0"
beautifulsoup4 = "^4.8.2"
pillow = "^7.0.0"
djangorestframework = "^3.11.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.4"
apscheduler = "^3.6.3"
boto3 = "^1.12.22"
django_storages = "^1.9.1"
django-secrets-manager = "^0.1.12"
django-allauth = "^0.41.0"
django-suit = "^0.2.28"
[tool.poetry.dev-dependencies]
django-extensions = "^2.2.8"
jupyter = "^1.0.0"
coverage = "^5.0.4"
pytest-django = "^3.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
